requirements.txt
aiogram==2.25.1
aiohttp==3.8.1
aiosignal==1.3.1
anyio==3.7.1
async-timeout==4.0.3
attrs==23.1.0
Babel==2.9.1
bcrypt==4.0.1
certifi==2023.7.22
charset-normalizer==2.1.1
click==8.1.6
ecdsa==0.18.0
environs==9.5.0
exceptiongroup==1.1.3
fastapi==0.78.0
frozenlist==1.4.0
greenlet==2.0.2
h11==0.14.0
httpcore==0.17.3
httpx==0.24.1
idna==3.4
magic-filter==1.0.11
marshmallow==3.20.1
multidict==6.0.4
packaging==23.1
passlib==1.7.4
pyasn1==0.5.0
pydantic==1.10.12
PyJWT==2.0.1
python-dotenv==1.0.0
python-jose==3.3.0
python-multipart==0.0.6
python-telegram-bot==20.4
pytz==2022.1
requests==2.31.0
rsa==4.9
six==1.16.0
sniffio==1.3.0
SQLAlchemy==1.4.37
starlette==0.19.1
typing_extensions==4.7.1
urllib3==2.0.4
uvicorn==0.18.1
uvloop==0.16.0
yarl==1.9.2
